Move scoring into its own module
